Extract the function angleBetween to a common math.js for reuse in the canvas rendering of the line offset
