+++ src/ol/math.js
@@ -212,3 +212,33 @@ export function wrap(n, min, max) {
   const range = max - min;
   return ((((n - min) % range) + range) % range) + min;
 }
+
+/**
+ * Compute the angle between p0pA and p0pB
+ * @param {import("./coordinate.js").Coordinate} p0 Point 0
+ * @param {import("./coordinate.js").Coordinate} pA Point A
+ * @param {import("./coordinate.js").Coordinate} pB Point B
+ * @return {number} a value in [0, 2PI]
+ */
+export function angleBetween(p0, pA, pB) {
+  const lenA = Math.sqrt(
+    (pA[0] - p0[0]) * (pA[0] - p0[0]) + (pA[1] - p0[1]) * (pA[1] - p0[1]),
+  );
+  const tangentA = [(pA[0] - p0[0]) / lenA, (pA[1] - p0[1]) / lenA];
+  const orthoA = [-tangentA[1], tangentA[0]];
+  const lenB = Math.sqrt(
+    (pB[0] - p0[0]) * (pB[0] - p0[0]) + (pB[1] - p0[1]) * (pB[1] - p0[1]),
+  );
+  const tangentB = [(pB[0] - p0[0]) / lenB, (pB[1] - p0[1]) / lenB];
+
+  // this angle can be clockwise or anticlockwise; hence the computation afterwards
+  let angle =
+    lenA === 0 || lenB === 0
+      ? 0
+      : Math.acos(
+          clamp(tangentB[0] * tangentA[0] + tangentB[1] * tangentA[1], -1, 1),
+        );
+  angle = Math.max(angle, 0.00001); // avoid a zero angle otherwise this is detected as a line cap
+  const isClockwise = tangentB[0] * orthoA[0] + tangentB[1] * orthoA[1] > 0;
+  return !isClockwise ? Math.PI * 2 - angle : angle;
+}
